requirements.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13.0
idna==3.3
isort==5.10.1
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySimpleGUI==4.60.0
PySocks==1.7.1
PyYAML==6.0
requests==2.27.1
selenium==4.1.5
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
